+++ journal.go
@@ -1,7 +1,7 @@
 package main
 
 type Journal struct {
-	Id         int16  `json:"id"`
+	Id         int32  `json:"id"`
 	Name       string `json:"name"`
 	Type       string `json:"type"` // S = Sale, P = Purchase, B = Bank, C = Cash, G = General
 	enterprise int32
